Before merging develop into nightly, sync `version.txt` with develop to (#80)
avoid a merge conflict; afterward, remove this sync commit.

The reason for removing the commit afterward is that we run
`update-version.sh`, which rewrites the file as necessary. Since
`nightly` merges from both `jdk21u` and `jdk21-dev`, but `develop`
only merges from jdk21u, version.txt may be newer in `nightly`. If we
sync `version.txt` with `develop` and don't remove it, as we do
currently, we avoid the merge conflict, but then `update-version.sh`
(which we treat as authoritative) overwrites `version.txt`, and we
have the two mechanisms writing two new commits each time the action
is run.